daemon/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pilot-daemon"
version = "0.3.0"
description = "AI-powered autonomous computer agent — full system control"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "websockets>=13.0",
    "pydantic>=2.6",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "tomli>=2.0; python_version < '3.12'",
    "tomli-w>=1.0",
    "cryptography>=43.0",
    "keyring>=25.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
full = [
    "chromadb>=0.5",
    "pyperclip>=1.8",
    "SecretStorage>=3.3; sys_platform == 'linux'",
]
llamacpp = ["llama-cpp-python>=0.3"]

# Tier 1: Game Changers
automation = [
    "pyautogui>=0.9",
    "Pillow>=10.0",
    "pyperclip>=1.8",
]
vision = [
    "easyocr>=1.7",
    "Pillow>=10.0",
]
browser = [
    "playwright>=1.40",
]

# Tier 2: Multipliers
documents = [
    "pypdf>=4.0",
    "python-docx>=1.0",
    "openpyxl>=3.1",
    "beautifulsoup4>=4.12",
]

# Everything
all = [
    "pilot-daemon[full,automation,vision,browser,documents]",
]

dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[project.scripts]
pilot-daemon = "pilot.server:main"

[tool.hatch.build.targets.wheel]
packages = ["pilot"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
